requirements.txt
django
djangorestframework
requests
django-cors-headers
pytest
channels
daphne